aspor-extraction-platform/requirements.txt
boto3==1.28.62
PyPDF2==3.0.1
python-docx==1.0.2
weasyprint==60.1
docx2python==2.0.1